rename tf1_model name
