+++ python/chronos/src/bigdl/chronos/model/tf1_model/_VanillaLSTM_keras.py
