+++ python/chronos/test/bigdl/chronos/model/tf1_model/test_VanillaLSTM_keras_tf1.py
